split plugins_common under separate internal libs
