+++ core/services/ocr2/plugins/ccip/commit_reporting_plugin.go
@@ -25,10 +25,12 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/abihelpers"
 	ccipconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/config"
+	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipcalc"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/contractutil"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/logpollerutil"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/pricegetter"
+	"github.com/smartcontractkit/chainlink/v2/core/utils/mathutil"
 
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/cache"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/hashlib"
@@ -260,7 +262,7 @@ func (r *CommitReportingPlugin) calculateMinMaxSequenceNumbers(ctx context.Conte
 		// missing a single, different log each, they would still be able to produce a valid report.
 		lggr.Warnf("Missing sequence number range [%d-%d]", nextInflightMin, minSeqNr)
 	}
-	if !contiguousReqs(lggr, minSeqNr, maxSeqNr, seqNrs) {
+	if !ccipcalc.ContiguousReqs(lggr, minSeqNr, maxSeqNr, seqNrs) {
 		return 0, 0, errors.New("unexpected gap in seq nums")
 	}
 	return minSeqNr, maxSeqNr, nil
@@ -289,7 +291,7 @@ func (r *CommitReportingPlugin) nextMinSeqNum(ctx context.Context, lggr logger.L
 		r.inflightReports.reset(lggr)
 		return nextMinOnChain, nextMinOnChain, nil
 	}
-	return max(nextMinOnChain, maxInflight+1), nextMinOnChain, nil
+	return mathutil.Max(nextMinOnChain, maxInflight+1), nextMinOnChain, nil
 }
 
 // All prices are USD ($1=1e18) denominated. We only generate prices we think should be updated;
@@ -351,7 +353,7 @@ func (r *CommitReportingPlugin) generatePriceUpdates(
 	if gasPrice == nil {
 		return nil, nil, fmt.Errorf("missing gas price %+v", sourceGasPriceWei)
 	}
-	sourceGasPriceUSD = calculateUsdPerUnitGas(gasPrice, sourceNativePriceUSD)
+	sourceGasPriceUSD = ccipcalc.CalculateUsdPerUnitGas(gasPrice, sourceNativePriceUSD)
 
 	lggr.Infow("Observing gas price", "observedGasPriceWei", gasPrice, "observedGasPriceUSD", sourceGasPriceUSD)
 	lggr.Infow("Observing token prices", "tokenPrices", tokenPricesUSD, "sourceNativePriceUSD", sourceNativePriceUSD)
@@ -598,12 +600,12 @@ func (r *CommitReportingPlugin) calculatePriceUpdates(observations []CommitObser
 		if len(tokenPriceObservations) <= r.F {
 			continue
 		}
-		medianPrice := median(tokenPriceObservations)
+		medianPrice := ccipcalc.BigIntMedian(tokenPriceObservations)
 
 		latestTokenPrice, exists := latestTokenPrices[token]
 		if exists {
 			tokenPriceUpdatedRecently := time.Since(latestTokenPrice.timestamp) < r.offchainConfig.FeeUpdateHeartBeat.Duration()
-			tokenPriceNotChanged := !deviates(medianPrice, latestTokenPrice.value, int64(r.offchainConfig.FeeUpdateDeviationPPB))
+			tokenPriceNotChanged := !ccipcalc.Deviates(medianPrice, latestTokenPrice.value, int64(r.offchainConfig.FeeUpdateDeviationPPB))
 			if tokenPriceUpdatedRecently && tokenPriceNotChanged {
 				r.lggr.Debugw("price was updated recently, skipping the update",
 					"token", token, "newPrice", medianPrice, "existingPrice", latestTokenPrice.value)
@@ -626,12 +628,12 @@ func (r *CommitReportingPlugin) calculatePriceUpdates(observations []CommitObser
 	destChainSelector := uint64(0)
 
 	if len(sourceGasObservations) > r.F {
-		usdPerUnitGas = median(sourceGasObservations)    // Compute the median price
-		destChainSelector = r.config.sourceChainSelector // Assuming plugin lane is A->B, we write to B the gas price of A
+		usdPerUnitGas = ccipcalc.BigIntMedian(sourceGasObservations) // Compute the median price
+		destChainSelector = r.config.sourceChainSelector             // Assuming plugin lane is A->B, we write to B the gas price of A
 
 		if latestGasPrice.value != nil {
 			gasPriceUpdatedRecently := time.Since(latestGasPrice.timestamp) < r.offchainConfig.FeeUpdateHeartBeat.Duration()
-			gasPriceNotChanged := !deviates(usdPerUnitGas, latestGasPrice.value, int64(r.offchainConfig.FeeUpdateDeviationPPB))
+			gasPriceNotChanged := !ccipcalc.Deviates(usdPerUnitGas, latestGasPrice.value, int64(r.offchainConfig.FeeUpdateDeviationPPB))
 			if gasPriceUpdatedRecently && gasPriceNotChanged {
 				usdPerUnitGas = big.NewInt(0)
 				destChainSelector = uint64(0)
@@ -670,7 +672,7 @@ func (r *CommitReportingPlugin) buildReport(ctx context.Context, lggr logger.Log
 		return commit_store.CommitStoreCommitReport{}, err
 	}
 
-	leaves, err := leavesFromIntervals(lggr, interval, r.config.leafHasher, sendRequests)
+	leaves, err := hashlib.LeavesFromIntervals(lggr, interval, r.config.leafHasher, sendRequests)
 	if err != nil {
 		return commit_store.CommitStoreCommitReport{}, err
 	}
@@ -719,7 +721,7 @@ func (r *CommitReportingPlugin) ShouldAcceptFinalizedReport(ctx context.Context,
 		return false, nil
 	}
 
-	epochAndRound := mergeEpochAndRound(reportTimestamp.Epoch, reportTimestamp.Round)
+	epochAndRound := ccipcalc.MergeEpochAndRound(reportTimestamp.Epoch, reportTimestamp.Round)
 	if err := r.inflightReports.add(lggr, parsedReport, epochAndRound); err != nil {
 		return false, err
 	}
@@ -788,7 +790,7 @@ func (r *CommitReportingPlugin) isStaleReport(ctx context.Context, lggr logger.L
 		return true
 	}
 
-	thisEpochAndRound := mergeEpochAndRound(reportTimestamp.Epoch, reportTimestamp.Round)
+	thisEpochAndRound := ccipcalc.MergeEpochAndRound(reportTimestamp.Epoch, reportTimestamp.Round)
 	return lastPriceEpochAndRound >= thisEpochAndRound
 }
 
@@ -825,7 +827,7 @@ func (r *CommitReportingPlugin) isStaleGasPrice(ctx context.Context, lggr logger
 		return true
 	}
 
-	if gasPriceUpdate.value != nil && !deviates(priceUpdates.UsdPerUnitGas, gasPriceUpdate.value, int64(r.offchainConfig.FeeUpdateDeviationPPB)) {
+	if gasPriceUpdate.value != nil && !ccipcalc.Deviates(priceUpdates.UsdPerUnitGas, gasPriceUpdate.value, int64(r.offchainConfig.FeeUpdateDeviationPPB)) {
 		lggr.Infow("Report is stale because of gas price",
 			"latestGasPriceUpdate", gasPriceUpdate.value,
 			"usdPerUnitGas", priceUpdates.UsdPerUnitGas,
@@ -846,7 +848,7 @@ func (r *CommitReportingPlugin) isStaleTokenPrices(ctx context.Context, lggr log
 
 	for _, tokenUpdate := range priceUpdates {
 		latestUpdate, ok := latestTokenPriceUpdates[tokenUpdate.SourceToken]
-		priceEqual := ok && !deviates(tokenUpdate.UsdPerToken, latestUpdate.value, int64(r.offchainConfig.FeeUpdateDeviationPPB))
+		priceEqual := ok && !ccipcalc.Deviates(tokenUpdate.UsdPerToken, latestUpdate.value, int64(r.offchainConfig.FeeUpdateDeviationPPB))
 
 		if !priceEqual {
 			lggr.Infow("Found non-stale token price", "token", tokenUpdate.SourceToken, "usdPerToken", tokenUpdate.UsdPerToken, "latestUpdate", latestUpdate.value)
