+++ core/services/ocr2/plugins/ccip/internal/ccipcalc/calc.go
@@ -1,0 +1,52 @@
+package ccipcalc
+
+import (
+	"math/big"
+	"sort"
+
+	"github.com/smartcontractkit/chainlink/v2/core/logger"
+)
+
+func ContiguousReqs(lggr logger.Logger, min, max uint64, seqNrs []uint64) bool {
+	if int(max-min+1) != len(seqNrs) {
+		return false
+	}
+	for i, j := min, 0; i <= max && j < len(seqNrs); i, j = i+1, j+1 {
+		if seqNrs[j] != i {
+			lggr.Errorw("unexpected gap in seq nums", "seqNr", i, "minSeqNr", min, "maxSeqNr", max)
+			return false
+		}
+	}
+	return true
+}
+
+func CalculateUsdPerUnitGas(sourceGasPrice *big.Int, usdPerFeeCoin *big.Int) *big.Int {
+	// (wei / gas) * (usd / eth) * (1 eth / 1e18 wei)  = usd/gas
+	tmp := new(big.Int).Mul(sourceGasPrice, usdPerFeeCoin)
+	return tmp.Div(tmp, big.NewInt(1e18))
+}
+
+func BigIntMedian(vals []*big.Int) *big.Int {
+	valsCopy := make([]*big.Int, len(vals))
+	copy(valsCopy[:], vals[:])
+	sort.Slice(valsCopy, func(i, j int) bool {
+		return valsCopy[i].Cmp(valsCopy[j]) == -1
+	})
+	return valsCopy[len(valsCopy)/2]
+}
+
+// deviation_parts_per_billion = ((x2 - x1) / x1) * 1e9
+func Deviates(x1, x2 *big.Int, ppb int64) bool {
+	// if x1 == 0, deviates if x2 != x1, to avoid the relative division by 0 error
+	if x1.BitLen() == 0 {
+		return x1.Cmp(x2) != 0
+	}
+	diff := big.NewInt(0).Sub(x1, x2)
+	diff.Mul(diff, big.NewInt(1e9))
+	diff.Div(diff, x1)
+	return diff.CmpAbs(big.NewInt(ppb)) > 0
+}
+
+func MergeEpochAndRound(epoch uint32, round uint8) uint64 {
+	return uint64(epoch)<<8 + uint64(round)
+}
