+++ core/services/ocr2/plugins/ccip/vars.go
@@ -1,0 +1,14 @@
+package ccip
+
+import (
+	"github.com/pkg/errors"
+)
+
+const (
+	MaxQueryLength       = 0       // empty for both plugins
+	MaxObservationLength = 250_000 // plugins's Observation should make sure to cap to this limit
+	CommitPluginLabel    = "commit"
+	ExecPluginLabel      = "exec"
+)
+
+var ErrCommitStoreIsDown = errors.New("commitStore is down")
